refactor(dashboard): move legacy lens converter lookup beside lens panel types
Keep firstLensVisConverterForChartBlocks next to lensVisConverters consumers
(models_lens_panel) pending Section 4 registry cleanup.
